Streamlining type and record management.
