+++ python/pcc.egg-info/PKG-INFO
@@ -1,6 +1,6 @@
 Metadata-Version: 1.1
 Name: pcc
-Version: 0.0.12
+Version: 0.0.13
 Summary: Predicate collection classes
 Home-page: https://github.com/Mondego/pcc
 Author: Rohan Achar
