+++ python/pcc/dataframe/cache_manager.py
@@ -1,0 +1,31 @@
+#################################################
+#### Attached Dataframe Stuff ###################
+#################################################
+
+class CacheManager(object):
+    def __init__(self):
+        
+        self.df_to_tp = dict()
+
+        self.attached_dataframes = set()
+
+        self.tp_to_attached_df = dict()
+
+    #################################################
+    ### Static Methods ##############################
+    #################################################
+    
+    
+    
+    #################################################
+    ### API Methods #################################
+    #################################################
+        
+    def add_records(self, records):
+        pass
+
+    #################################################
+    ### Private Methods #############################
+    #################################################
+    
+
